add typings test for spread object with unexpected column @ update.
