+++ test/typings/test-d/index.test-d.ts
@@ -783,36 +783,6 @@ async function testReturning(db: Kysely<Database>) {
   expectError(db.insertInto('person').values(person).returning('not_column'))
 }
 
-async function testUpdate(db: Kysely<Database>) {
-  const r1 = await db
-    .updateTable('pet as p')
-    .where('p.id', '=', '1')
-    .set({ name: 'Fluffy' })
-    .executeTakeFirst()
-
-  expectType<UpdateResult>(r1)
-
-  // Non-existent column
-  expectError(
-    db
-      .updateTable('pet as p')
-      .where('p.id', '=', '1')
-      .set({ not_a_column: 'Fluffy' })
-  )
-
-  // GeneratedAlways column is not allowed to be updated
-  expectError(db.updateTable('book').set({ id: 1, name: 'foo' }))
-
-  db.updateTable('book').set({ name: 'bar' })
-
-  // Nullable column as undefined
-  const mutationObject: { last_name: string | undefined } = {
-    last_name: 'smith',
-  }
-
-  db.updateTable('person').set(mutationObject)
-}
-
 async function testDelete(db: Kysely<Database>) {
   const r1 = await db.deleteFrom('pet').where('id', '=', '1').executeTakeFirst()
   expectType<DeleteResult>(r1)
